- Enable 2.6.6-rc patches. - Enable jfs dmapi patches (disable in config files).
suse-commit: 22afee0e8cac78e954b266e4358dc5ae66540685

+++ fs/xfs/linux/xfs_aops.c
@@ -1016,7 +1016,7 @@ linvfs_get_blocks_direct(
 					create, 1, BMAPI_WRITE|BMAPI_DIRECT);
 }
 
-STATIC int
+STATIC ssize_t
 linvfs_direct_IO(
 	int			rw,
 	struct kiocb		*iocb,
